Add DSP tests (for SiPM)
