+++ tests/dsp/test_build_dsp.py
@@ -5,31 +5,8 @@ import pytest
 from pygama import lgdo
 from pygama.dsp import build_dsp
 from pygama.lgdo.lh5_store import LH5Store, ls
-from pygama.raw import build_raw
 
 config_dir = Path(__file__).parent / "configs"
-
-
-@pytest.fixture(scope="module")
-def multich_raw_file(lgnd_test_data):
-    out_file = "/tmp/L200-comm-20211130-phy-spms.lh5"
-    out_spec = {
-        "FCEventDecoder": {
-            "ch{key}": {
-                "key_list": [[0, 6]],
-                "out_stream": out_file + ":{name}",
-                "out_name": "raw",
-            }
-        }
-    }
-
-    build_raw(
-        in_stream=lgnd_test_data.get_path("fcio/L200-comm-20211130-phy-spms.fcio"),
-        out_spec=out_spec,
-        overwrite=True,
-    )
-
-    return out_file
 
 
 def test_build_dsp_basics(lgnd_test_data, dsp_test_file):
@@ -49,30 +26,14 @@ def test_build_dsp_basics(lgnd_test_data, dsp_test_file):
         )
 
 
-def test_build_dsp_channelwise(multich_raw_file):
-    chan_config = {
-        "ch0/raw": f"{config_dir}/sipm-dsp-config.json",
-        "ch1/raw": f"{config_dir}/sipm-dsp-config.json",
-        "ch2/raw": f"{config_dir}/sipm-dsp-config.json",
-    }
-
-    out_file = "/tmp/L200-comm-20211130-phy-spms_dsp.lh5"
-    build_dsp(
-        multich_raw_file,
-        out_file,
-        {},
-        n_max=5,
-        lh5_tables=chan_config.keys(),
-        chan_config=chan_config,
-        write_mode="r",
-    )
-
-    assert ls(out_file) == ["ch0", "ch1", "ch2", "dsp_info"]
-    assert ls(out_file, "ch0/") == ["ch0/dsp"]
-    assert ls(out_file, "ch0/dsp/") == ["ch0/dsp/energies", "ch0/dsp/trigger_pos"]
+def test_build_dsp_spms_channelwise(dsp_test_file_spm):
+    
+    assert ls(dsp_test_file_spm) == ["ch0", "ch1", "ch2", "dsp_info"]
+    assert ls(dsp_test_file_spm, "ch0/") == ["ch0/dsp"]
+    assert ls(dsp_test_file_spm, "ch0/dsp/") == ["ch0/dsp/energies", "ch0/dsp/trigger_pos"]
 
     store = LH5Store()
-    lh5_obj, n_rows = store.read_object("/ch0/dsp/energies", out_file)
+    lh5_obj, n_rows = store.read_object("/ch0/dsp/energies", dsp_test_file_spm)
     assert isinstance(lh5_obj, lgdo.ArrayOfEqualSizedArrays)
     assert len(lh5_obj) == 5
     assert len(lh5_obj.nda[0]) == 20
